requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
pandas==1.4.2
sentence-transformers==2.2.2
faiss-cpu==1.7.3
numpy==1.22.4
